Massively restructured things, file structure should be clearer now
