+++ exec/bochsrc.txt
@@ -3,8 +3,8 @@ romimage:        file=/usr/share/bochs/BIOS-bochs-latest
 vgaromimage:     file=/usr/share/bochs/VGABIOS-lgpl-latest
 ata0-master:     type=cdrom, path=ShantOS.iso, status=inserted
 boot:            cdrom
-log:             bochslog.txt
+log:             log/bochslog.txt
 clock:           sync=realtime, time0=local
 cpu:             count=1, ips=1000000
-com1: enabled=1, mode=file, dev=com1.out
+com1: enabled=1, mode=file, dev=log/com1.out
