+++ exec/link.ld
@@ -24,3 +24,4 @@ SECTIONS {
         *(.bss)              /* all bss sections from all files */
     }
 }
+
